# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger SaperX test",
        "description": "API para o teste da SaperX",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "marcosvm000@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/phonebook": {
            "get": {
                "tags": [
                    "Phonebook"
                ],
                "summary": "Gera o relatório dos nomes e números da agenda",
                "operationId": "listPhonebookContacts",
                "responses": {
                    "200": {
                        "description": "List of contacts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Example name"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "example@example"
                                                    },
                                                    "birthdate": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "yyyy-mm-dd"
                                                    },
                                                    "CPF": {
                                                        "type": "string",
                                                        "example": "99999999988"
                                                    },
                                                    "phones": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Phonebook"
                ],
                "summary": "Cria um nome e número ná agenda",
                "operationId": "createPhonebookContact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "YYYY-MM-DD"
                                    },
                                    "CPF": {
                                        "description": "Brazilian CPF number",
                                        "type": "string",
                                        "pattern": "^\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}$"
                                    },
                                    "phones": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "telefone_com_ddd"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phonebook contact created!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Phonebook created!",
                                    "data": {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "birthdate": "1990-01-01",
                                        "CPF": "123.456.789-00",
                                        "phones": [
                                            "(123) 456-7890"
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Something is wrong!"
                    },
                    "422": {
                        "description": "Data invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Data invalid",
                                    "errors": {
                                        "name": [
                                            "The name field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/phonebook/{id}": {
            "put": {
                "tags": [
                    "Phonebook"
                ],
                "summary": "Edita algum campo do cadastro da agenda",
                "operationId": "updatePhonebookContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the phonebook contact",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "YYYY-MM-DD"
                                    },
                                    "CPF": {
                                        "description": "Brazilian CPF number",
                                        "type": "string",
                                        "pattern": "^\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}$"
                                    },
                                    "phones": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "telefone_com_ddd"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phonebook contact updated!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Phone number updated!",
                                    "data": {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "birthdate": "1990-01-01",
                                        "CPF": "123.456.789-00",
                                        "phones": [
                                            "(123) 456-7890"
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Something is wrong!"
                    },
                    "404": {
                        "description": "Phone number not found!"
                    },
                    "422": {
                        "description": "Data invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Data invalid",
                                    "errors": {
                                        "name": [
                                            "The name field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Phonebook"
                ],
                "summary": "Deleta um registro da agenda pelo id informado",
                "operationId": "deletePhonebookContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the phonebook contact",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Phone number deleted!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Phone number deleted!"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Phone number not deleted!"
                    },
                    "404": {
                        "description": "Phone number not found!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}